pyproject.toml
[tool.poetry]
name = "insight-app"
version = "0.1.0"
description = "Insight application"
authors = ["Abubakr <abubakr.max@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"

# Add your app dependencies here:
flask = "^3.0"
psycopg2-binary = "^2.9.11"
asyncpg = "^0.30.0"
fastapi = "^0.119.1"
uvicorn = "^0.38.0"
alembic = "^1.17.0"
python-dotenv = "^1.2.1"
python-jose = "^3.5.0"
pydantic = {extras = ["email"], version = "^2.12.5"}

[tool.poetry.dev-dependencies]
pytest = "^8.0"
black = "^24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
